JobsScraper/flink/requirements.txt
#pyflink==1.0
python-dotenv==1.0.0
beautifulsoup4==4.11.1
pydantic==1.10.11
spacy==3.6.1
locationtagger==0.0.1
unidecode==1.3.7
#db-sqlite3
deep-translator